🐛 Fix:#20-수정사항 반영
